Allow TypeScript's predefined-type keywords to update-match type_identifier
typescript-add-generics: swapping a primitive type for a generic type
parameter (private value: number -> private value: T) always fell out as
Delete + Insert, because kinds_update_allowed had no cross-kind exception
for predefined_type's keyword leaves (number, string, boolean, ...)
against type_identifier - exactly the "add generics" edit this fixture is
named for. Adding predefined_type itself to the family (rather than its
keyword children) was tried first and made it worse (14 -> 18 mismatches):
predefined_type is the keyword leaf's parent, so APTED took a same-cost
but structurally wrong parent-level match instead of deleting the wrapper
and matching the keyword leaf, which is what the human mapping wants.
Restricting the new family to the keyword leaves fixes it cleanly:
14 -> 10 mismatches on this fixture, 0 regressions across the other 510
solved fixtures (research/data/quality/optimal_solutions_benchmark.csv,
sum 7358 -> 7354).

Adding a 9th operator family overflowed the u8 bitmask
(`nodes::operator_family_mask`/`language_operator_family_mask`,
`KindCostClass::operator_families`, `UnitCostModel::language_family_mask`):
1u8 << 8 wraps to bit 0 in a release build, silently colliding the new
family with COMPARISON_OPS. Caught by the existing
operator_family_masks_agree_with_string_scanning_kinds_update_allowed
consistency test, not by inspection. Widened all four to u16.

+++ src/code.rs
@@ -383,7 +383,7 @@ pub struct KindCostClass {
     pub literal_like: bool,
     /// Bit `i` set iff this kind is in `ALL_OPERATOR_FAMILIES[i]` (see
     /// [`crate::diff::nodes::operator_family_mask`]). A kind can be in several families at once.
-    pub operator_families: u8,
+    pub operator_families: u16,
 }
 
 /**
